renamed functions to utils
